:recycle: move actions related to delegate to account
